improve performance; fix ignored patience parameter
The evaluation of the parameters is now done in one JS function call per iteration, instead of *num_points* calls per iteration. Major (~2x) speedup for large datasets.

There was also a mismatch in the number of arguments in the `crwap` call that resulted in the last paramter (`patience`) being ignored.